Chapter07/requirements.txt
absl-py==1.4.0
aiohttp==3.8.4
aiosignal==1.3.1
alembic==1.10.2
antlr4-python3-runtime==4.9.3
async-timeout==4.0.2
attrs==22.2.0
boto3==1.26.98
botocore==1.29.98
cachetools==5.3.0
category-encoders==2.5.1.post0
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.0
cloudpickle==2.2.1
cmake==3.26.0
contourpy==1.0.7
cycler==0.11.0
databricks-cli==0.17.6
datasets==2.10.1
dill==0.3.6
docker==6.0.1
docstring-parser==0.15
einops==0.6.0
entrypoints==0.4
filelock==3.10.2
flask==2.2.3
fonttools==4.39.2
frozenlist==1.3.3
fsspec==2023.3.0
ftfy==6.1.1
gitdb==4.0.10
gitpython==3.1.31
google-auth==2.16.2
google-auth-oauthlib==0.4.6
greenlet==2.0.2
grpcio==1.51.3
gunicorn==20.1.0
huggingface-hub==0.13.3
idna==3.4
importlib-resources==5.12.0
ipywidgets==8.0.5
itsdangerous==2.1.2
jinja2==3.1.2
jmespath==1.0.1
joblib==1.2.0
jsonargparse==4.9.0
jupyterlab-widgets==3.0.6
kiwisolver==1.4.4
lightning-flash==0.8.1.post0
lightning-utilities==0.8.0
lit==16.0.0
llvmlite==0.39.1
mako==1.2.4
markdown==3.4.2
markdown-it-py==2.2.0
markupsafe==2.1.2
matplotlib==3.7.1
mdurl==0.1.2
mlflow==2.2.2
mpmath==1.3.0
multidict==6.0.4
multiprocess==0.70.14
networkx==3.0
nltk==3.8.1
numba==0.56.4
numpy==1.23.5
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.2.10.91
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusparse-cu11==11.7.4.91
nvidia-nccl-cu11==2.14.3
nvidia-nvtx-cu11==11.7.91
oauthlib==3.2.2
omegaconf==2.3.0
pandas==1.5.3
patsy==0.5.3
pillow==9.4.0
protobuf==3.19.6
pyarrow==11.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pydeprecate==0.3.2
pyjwt==2.6.0
pyparsing==3.0.9
pytorch-lightning==1.8.6
pytorch-tabnet==4.0
pytz==2022.7.1
pyyaml==5.4.1
querystring-parser==1.2.4
regex==2023.3.22
requests==2.28.2
requests-oauthlib==1.3.1
responses==0.18.0
rich==13.3.2
rsa==4.9
s3transfer==0.6.0
scikit-learn==1.2.2
scipy==1.8.0
sentence-transformers==2.2.2
sentencepiece==0.1.97
setuptools==59.5.0
shap==0.41.0
slicer==0.0.7
smmap==5.0.0
sqlalchemy==2.0.7
sqlparse==0.4.3
statsmodels==0.13.5
sympy==1.11.1
tabulate==0.9.0
tensorboard==2.12.0
tensorboard-data-server==0.7.0
tensorboard-plugin-wit==1.8.1
tensorboardx==2.6
threadpoolctl==3.1.0
tokenizers==0.13.2
torch==1.9.0
torchmetrics==0.10.3
torchvision==0.10.0
tqdm==4.65.0
transformers==4.27.2
triton==2.0.0
urllib3==1.26.15
websocket-client==1.5.1
werkzeug==2.2.3
widgetsnbextension==4.0.6
xxhash==3.2.0
yarl==1.8.2
ray[tune]==2.3.0
optuna==3.0.4